+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## fugit 1.1.7  not yet released
 
-* Prevent 0 as day of month, gh-10
+* Prevent 0 as a month in crons, gh-10
+* Prevent 0 as a day of month in crons, gh-10
 
 
 ## fugit 1.1.6  released 2018-09-05